+++ src/components/numeral-date/numeral-date-test.stories.tsx
@@ -14,7 +14,6 @@ import CarbonProvider from "../carbon-provider";
 
 export default {
   title: "Numeral Date/Test",
-  excludeStories: ["NumeralDateComponent"],
   parameters: {
     info: { disable: true },
     chromatic: {
@@ -212,20 +211,3 @@ export const TooltipPosition = () => {
 };
 
 TooltipPosition.storyName = "tooltip position";
-
-export const NumeralDateComponent = (props: Partial<NumeralDateProps>) => {
-  const [value, setValue] = React.useState({
-    dd: "",
-    mm: "",
-    yyyy: "",
-  });
-
-  return (
-    <NumeralDate
-      value={value}
-      onChange={(e) => setValue(e.target.value)}
-      label="Default"
-      {...props}
-    />
-  );
-};