+++ src/components/numeral-date/numeral-date.spec.tsx
@@ -7,7 +7,8 @@ import NumeralDate, {
   ValidDateFormat,
 } from "./numeral-date.component";
 import Textbox from "../textbox";
-import { StyledDateField, StyledNumeralDate } from "./numeral-date.style";
+import Typography from "../typography";
+import { StyledNumeralDate } from "./numeral-date.style";
 import {
   assertStyleMatch,
   testStyledSystemMargin,
@@ -21,7 +22,6 @@ import CarbonProvider from "../carbon-provider/carbon-provider.component";
 import { ErrorBorder, StyledHintText } from "../textbox/textbox.style";
 import StyledValidationMessage from "../../__internal__/validation-message/validation-message.style";
 import Logger from "../../__internal__/utils/logger";
-import StyledInputPresentation from "../../__internal__/input/input-presentation.style";
 
 jest.mock("../../__internal__/utils/logger");
 
@@ -117,7 +117,7 @@ describe("NumeralDate", () => {
       const expected =
         "Forbidden prop dateFormat supplied to NumeralDate. " +
         "Only one of these date formats is allowed: " +
-        "['dd', 'mm', 'yyyy'], ['mm', 'dd', 'yyyy'], ['dd', 'mm'], ['mm', 'dd'], ['mm', 'yyyy']";
+        "['dd', 'mm', 'yyyy'], ['mm', 'dd', 'yyyy'], ['yyyy', 'mm', 'dd'], ['dd', 'mm'], ['mm', 'dd'], ['mm', 'yyyy']";
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore:next-line testing incorrect date format
@@ -196,7 +196,6 @@ describe("NumeralDate", () => {
       assertStyleMatch(
         {
           display: "inline-flex",
-          height: "40px",
           fontSize: "14px",
           fontWeight: "400",
         },
@@ -378,9 +377,10 @@ describe("NumeralDate", () => {
 
   it("has proper default dateFormat prop", () => {
     wrapper = renderWrapper({ dateFormat: undefined });
-    expect(wrapper.find(Textbox).at(0).props().placeholder).toEqual("dd");
-    expect(wrapper.find(Textbox).at(1).props().placeholder).toEqual("mm");
-    expect(wrapper.find(Textbox).at(2).props().placeholder).toEqual("yyyy");
+
+    expect(wrapper.find(Typography).at(0).text()).toBe("Day");
+    expect(wrapper.find(Typography).at(1).text()).toBe("Month");
+    expect(wrapper.find(Typography).at(2).text()).toBe("Year");
   });
 
   describe("Clicking off the component", () => {
@@ -462,7 +462,7 @@ describe("NumeralDate", () => {
       const input = wrapper.find("input").at(0);
       const event = { key: "1", preventDefault: preventDefaultMock };
       act(() => {
-        input.simulate("keypress", event);
+        input.simulate("keydown", event);
       });
       expect(preventDefaultMock).not.toHaveBeenCalled();
     });
@@ -471,7 +471,7 @@ describe("NumeralDate", () => {
       const input = wrapper.find("input").at(0);
       const event = { key: key[0], preventDefault: preventDefaultMock };
       act(() => {
-        input.simulate("keypress", event);
+        input.simulate("keydown", event);
       });
       expect(preventDefaultMock).toHaveBeenCalled();
     });
@@ -544,15 +544,12 @@ describe("NumeralDate", () => {
   });
 
   describe("new validation design", () => {
-    describe("hint text/labelHelp", () => {
-      it("is visible when the prop is passed", () => {
-        wrapper = renderNewValidationsWrapper({
-          labelHelp: "Example hint text",
-        });
-        expect(wrapper.find(StyledHintText).text()).toEqual(
-          "Example hint text"
-        );
+    it("hint text is visible and tooltip is not rendered when the labelHelp prop is passed", () => {
+      wrapper = renderNewValidationsWrapper({
+        labelHelp: "Example hint text",
       });
+      expect(wrapper.find(StyledHintText).text()).toEqual("Example hint text");
+      expect(wrapper.find(StyledHelp).exists()).toEqual(false);
     });
 
     describe("on error", () => {
@@ -690,26 +687,5 @@ describe("NumeralDate", () => {
         expect(ref.current).toBe(null);
       });
     });
-  });
-
-  it("renders with the expected border radius styling on first and last inputs", () => {
-    wrapper = renderWrapper({ value: { dd: "02", mm: "01", yyyy: "2020" } });
-    assertStyleMatch(
-      {
-        borderTopLeftRadius: "var(--borderRadius050)",
-        borderBottomLeftRadius: "var(--borderRadius050)",
-      },
-      wrapper.find(StyledDateField).first(),
-      { modifier: `${StyledInputPresentation}` }
-    );
-
-    assertStyleMatch(
-      {
-        borderTopRightRadius: "var(--borderRadius050)",
-        borderBottomRightRadius: "var(--borderRadius050)",
-      },
-      wrapper.find(StyledDateField).last(),
-      { modifier: `${StyledInputPresentation}` }
-    );
   });
 });