+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [124.6.0](https://github.com/Sage/carbon/compare/v124.5.0...v124.6.0) (2024-01-05)
+
+
+### Features
+
+* **progress-tracker:** add inline label and label width prop ([5742c81](https://github.com/Sage/carbon/commit/5742c8199cbabe91fc4f561aba232d46dad3a4b1))
+
+## [124.5.0](https://github.com/Sage/carbon/compare/v124.4.1...v124.5.0) (2024-01-04)
+
+
+### Features
+
+* **numeral-date:** add date format and update to align with DS ([a292c86](https://github.com/Sage/carbon/commit/a292c86cc0eda7b363e545965b7fdc39b5ddfd1c))
+
 ### [124.4.1](https://github.com/Sage/carbon/compare/v124.4.0...v124.4.1) (2023-12-19)